site/requirements.txt
python-dotenv
flask-sqlalchemy
flask-migrate
flask-wtf
flask-login
flask-rbac
flask_uploads
ffmpeg-python
opencv-python
pyee
Flask-Mail
raven
mysqlclient
gunicorn
gevent
